Make dock dots scalable
This patch makes the three dots in the dock change their size when bigger
icon size is used. The logic here is such that the dots are scaled only when
a certain icon size is reached; for instance, the dots will remain unchanged
if icon size is less than 128x128 px, and will be twice as big if it's set to
a value from 128x128 to 184x184 px. The get_dot_mult() func calculates and
returns the scale multiplier. In the draw_dot() func, XDrawLine() and
XDrawPoint() were replaced with XFillRectangle() because those funcs do not
allow their result to be scalable. This patch does not require additional
icon size values (those bigger than 96x96 px) to exist, but makes no use
without them.