Add icon sizes
Original patch of: Xander <charlesdefaux@gmail.com>

The classic WindowMaker allows tile icon size up to 96x96 px.
This patch rises this value to 256x256 px. It also sets the max number
of visible items in the corresponding pref list to 29 (instead of 9),
so the new values show up in the list. This patch makes it possible
to use dockapps that allow changing their size to bigger than 64x64 px, see
https://gitlab.com/xander1988/dockapps for more info.